+++ packages/aws-lambda/src/wrapper.js
@@ -257,13 +257,7 @@ function postHandler(entrySpan, error, result, callback) {
     plugins: [{ name: 'com.instana.plugin.aws.lambda', entityId: identityProvider.getEntityId(), data: metricsData }]
   };
 
-  // Sending data from an AWS Lambda is always a fire and forget operation. We do not wait for the response from the
-  // Instana back end. This reduces the delay we add to the runtime of a Lambda handler significantly. The drawback we
-  // need to accept for that reduced runtime penalty is that we cannot know if the data has reached the Instana back
-  // end or not. Thus, the callback we pass to sendBundle is never called with an error.
-  // Also, we would not want to propagate an error from the backend request anyway - the customer's lambda handler needs
-  // to finish successfully, no matter if we have been able to report metrics and spans.
-  backendConnector.sendBundle({ spans, metrics: metricsPayload }, callback);
+  backendConnector.sendBundle({ spans, metrics: metricsPayload }, true, callback);
 }
 
 exports.currentSpan = function getHandleForCurrentSpan() {